lambda: modularize - utils
